# faster computation of transaction ID and length