@pbennett 2021-01-31
Just in case it wasn't known, running benchmark tests with -benchmem flag will track memory allocations as well which would be a great comparison to include here (in the already great performance improvment).